requirements.txt
aiomysql
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
aiosqlite
python-dotenv
PyMySQL